refactor: logger middleware
