+++ server/src/middleware/logger.js
@@ -1,0 +1,52 @@
+// @ts-check
+/* eslint-disable prefer-rest-params */
+const bytes = require('bytes')
+
+const { log, HELPERS } = require('@rm/logger')
+
+/** @type {import('@rm/types').ExpressMiddleware} */
+function loggerMiddleware(req, res, next) {
+  const start = process.hrtime()
+
+  const oldWrite = res.write
+  const oldEnd = res.end
+  let resBodySize = 0
+
+  res.write = function write(chunk) {
+    resBodySize += chunk.length
+    oldWrite.apply(res, arguments)
+    return false
+  }
+
+  res.end = function end(chunk) {
+    if (chunk) {
+      resBodySize += chunk.length
+    }
+    oldEnd.apply(res, arguments)
+    return res
+  }
+
+  res.on('finish', () => {
+    const [seconds, nanoseconds] = process.hrtime(start)
+    const responseTime = (seconds * 1000 + nanoseconds / 1e6).toFixed(3) // in milliseconds
+    log.debug(
+      HELPERS.express,
+      req.method,
+      req.originalUrl,
+      HELPERS.statusCode(res.statusCode),
+      `${responseTime}ms`,
+      '|',
+      // @ts-ignore
+      HELPERS.download(bytes(req.bodySize)),
+      HELPERS.upload(bytes(resBodySize)),
+      '|',
+      req.user ? req.user.username : 'Not Logged In',
+      req.headers['x-forwarded-for']
+        ? `| ${req.headers['x-forwarded-for']}`
+        : '',
+    )
+  })
+  next()
+}
+
+module.exports = { loggerMiddleware }
